==> 04a_challenge1.ipynb <==
# 04a - Challenge 1

Welcome to this challenge lab! Use this notebook to guide you in building a solution to address the following challenge.

**Challenge:** Load the featurizer and credit risk models which you registered in the MLflow model registry (in the previous lab), and use them to perform pre-processing and make predictions for some sample input data.

### 1. Set up environment

Restore variables

%store -r train_data_path test_data_path
%store -r bucket_name model_prefix model_artifact
%store -r role

# Retrieve the mflow tracking server from the previous lab
%store -r mlflow_arn

# Retrieve the logged MLflow model objects from the previous lab
%store -r logged_featurizer_model logged_creditrisk_model

Import the necessary libraries:

import boto3
import mlflow
import xgboost
import pandas as pd
from time import gmtime, strftime

Set the MLflow tracking server

# Set the MLflow tracking server

mlflow.set_tracking_uri(mlflow_arn)

### 2. Load the featurizer model from the MLflow model registry

You have imported the stored value `logged_featurizer_model` from the previous lab. This contains an MLflow object of class ModelInfo. Use the [ModelInfo documentation](https://mlflow.org/docs/latest/api_reference/python_api/mlflow.models.html#mlflow.models.model.ModelInfo) to figure out how to obtain the model URI for this model.

Use this model URI to load the sklearn featurizer model from the MLflow model registry. For guidance, see the [MLflow sklearn load_model() docs](https://mlflow.org/docs/latest/api_reference/python_api/mlflow.sklearn.html#mlflow.sklearn.load_model). Alternatively, use the MLflow console from SageMaker Studio to navigate to this model, and view the model code examples.

Insert your code here:

# Load featurizer sklearn model from MLflow model registry

featurizer_model = 

Uncomment and run the below to see the solution:

# %load solutions/4a_s2.py

### 3. Pre-process input data

Define a function to pre-process the input data. This should receive a dataframe with the input data, and the featurizer mode to use. The output should be a dataframe with the pre-processed data.

Refer back to notebook 2 to recall how the data needs to be pre-processed.

Insert your code here:

# Pre-process the input

def preprocess(df, featurizer_model):


Uncomment and run the below to see the solution:

# %load solutions/4a_s3.py

Test your function to pre-process the input data as follows:

# retrieve the input data
df1 = pd.read_csv('inference.csv')
csv_data = df1.to_csv(index=False, header=False)

# preprocess the data
payload_input = preprocess(df1,featurizer_model)
print (payload_input)
print("Number of samples in the payload:",len(payload_input))

### 4. Load the XGBoost credit risk model from the MLflow model registry

You have imported the stored value `logged_creditrisk_model` from the previous lab. This contains an MLflow object of class ModelInfo. Use the [ModelInfo documentation](https://mlflow.org/docs/latest/api_reference/python_api/mlflow.models.html#mlflow.models.model.ModelInfo) to figure out how to obtain the model URI for this model.

Use this model URI to load the xgboost featurizer model from the MLflow model registry. For guidance, see the [MLflow xgboost load_model() docs](https://mlflow.org/docs/latest/api_reference/python_api/mlflow.xgboost.html). 

Insert your code here:

# Load xgboost model from MLflow model registry

creditrisk_model =

Uncomment and run the below to see the solution:

# %load solutions/4a_s4.py

### 5. Use the credit risk model with the pre-processed input data to predict credit risk scoress

In the final step, use the credit risk model to predict credit risk scores for the pre-processed input data.

Insert your code here:

# Run predictions on the pre-processed payload

predictions =
predictions

Uncomment and run the below to see the solution:

# %load solutions/4a_s5.py
